Relocate FilePicker UI to own file
